refactor: make cli api more reusable
