+++ cli/testdata/pwd_id_ed25519
